+++ lib/membrane/bin/pad_data.ex
@@ -8,7 +8,7 @@ defmodule Membrane.Bin.PadData do
     - `:name` - see `t:Membrane.Pad.name/0`. Do not mistake with `:ref`
     - `:options` - options passed in `Membrane.ChildrenSpec` when linking pad
     - `:ref` - see `t:Membrane.Pad.ref/0`
-    - `max_cardinality` - specyfies maximal possible number of instances of a dynamic pads that can occur within single element. `nil` for pads with `availability: :always`.
+    - `max_instances` - specifies maximal possible number of instances of a dynamic pad that can exist within single element. Equals `nil` for pads with `availability: :always`.
 
   Other fields in the struct ARE NOT PART OF THE PUBLIC API and should not be
   accessed or relied on.
@@ -24,7 +24,7 @@ defmodule Membrane.Bin.PadData do
           availability: Membrane.Pad.availability(),
           direction: Membrane.Pad.direction(),
           name: Membrane.Pad.name(),
-          max_cardinality: Membrane.Pad.max_cardinality() | nil,
+          max_instances: Membrane.Pad.max_instances() | nil,
           spec_ref: private_field,
           link_id: private_field,
           endpoint: private_field,
@@ -49,5 +49,5 @@ defmodule Membrane.Bin.PadData do
     :linked_in_spec?
   ]
 
-  defstruct @enforce_keys ++ [:max_cardinality]
+  defstruct @enforce_keys ++ [:max_instances]
 end