+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.2.0
- * Add `:max_cardinality` option for dynamic pads. [#876](https://github.com/membraneframework/membrane_core/pull/876)
+ * Add `:max_instances` option for dynamic pads. [#876](https://github.com/membraneframework/membrane_core/pull/876)
 
 ## 1.1.2
  * Remove 'failed to insert a metric' stalker warning. [#849](https://github.com/membraneframework/membrane_core/pull/849)